requirements.txt
Flask
requests
Pillow
spotipy
colorthief
colorgram.py